Move nested functions out of `disassemble_from_template_inner`
Use `case_names` when formatting booleans

Add lookup tables to some templates

Moved lookup tables to new file Lookups.cc and conversion functions to new file Conversions.cc for better re-use
Fix string encoding vers-resource
