+++ src/Conversions.hh
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <string>
+
+using namespace std;
+
+std::string decode_mac_roman(const char* data, size_t size);
+std::string decode_mac_roman(const std::string& data);
+std::string decode_mac_roman(char data);
